Optimized mongo query for readings-daily-aggregates retrieval.
The query was on the non-indexed properties `sensorId`, `day`, `source`
and `measurementType`. Since the `_id` is constructed by the aggregator
which writes the element to the collection concatenating those 4
properties, we can re-construct the `_id` using the same method and use
it for the query. This makes the query perform much better (no
full-table-scan involved) and avoids the burden of creating an index.